services/ingestion/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.6
pydantic-settings==2.8.1
python-multipart==0.0.20
pdfplumber==0.11.5
tabula-py==2.10.0
pandas==2.2.3

